+++ ChipDistributionGame/Shared/src/main/java/nl/han/asdb/shared/interfaces/eventmanager/events/game/SellChipsEvent.java
@@ -59,7 +59,7 @@ public class SellChipsEvent extends Event {
         return pricePerItem;
     }
 
-    public void setPricePerItem(short pricePerItem) {
+    public void setPricePerItem(double pricePerItem) {
         this.pricePerItem = pricePerItem;
     }
